test(use-async-callback): migrate tests to browser mode
